chore: refactor code for reuse for ComponentDefinition schema
