+++ packages/validators/src/schemas/v2023_09_28/experience.ts
@@ -1,10 +1,31 @@
 import { z } from 'zod';
 import { SchemaVersions } from '../schemaVersions';
-import {
-  DefinitionPropertyTypeSchema,
-  DefinitionPropertyKeySchema,
-  PrimitiveValueSchema,
-} from '../componentDefinition';
+
+export const DefinitionPropertyTypeSchema = z.enum([
+  'Text',
+  'RichText',
+  'Number',
+  'Date',
+  'Boolean',
+  'Location',
+  'Media',
+  'Object',
+  'Hyperlink',
+  'Array',
+  'Link',
+]);
+
+export const DefinitionPropertyKeySchema = z
+  .string()
+  .regex(/^[a-zA-Z0-9-_]{1,32}$/, { message: 'Property needs to match: /^[a-zA-Z0-9-_]{1,32}$/' });
+
+export const PrimitiveValueSchema = z.union([
+  z.string(),
+  z.boolean(),
+  z.number(),
+  z.record(z.any(), z.any()),
+  z.undefined(),
+]);
 
 const uuidKeySchema = z
   .string()
@@ -105,31 +126,35 @@ const ComponentTreeNodeSchema: z.ZodType<ComponentTreeNode> = BaseComponentTreeN
   children: z.lazy(() => ComponentTreeNodeSchema.array()),
 });
 
-const ComponentSettingsSchema = z.object({
-  variableDefinitions: z.record(
-    z.string().regex(/^[a-zA-Z0-9-_]{1,54}$/), // Here the key is <variableName>_<nanoidId> so we need to allow for a longer length
-    z.object({
-      displayName: z.string().optional(),
-      type: DefinitionPropertyTypeSchema,
-      defaultValue: PrimitiveValueSchema.or(ComponentPropertyValueSchema).optional(),
-      description: z.string().optional(),
-      group: z.string().optional(),
-      validations: z
-        .object({
-          required: z.boolean().optional(),
-          format: z.literal('URL').optional(),
-          in: z
-            .array(
-              z.object({
-                value: z.union([z.string(), z.number()]),
-                displayName: z.string().optional(),
-              }),
-            )
-            .optional(),
-        })
+export const ComponentVariableSchema = z.object({
+  displayName: z.string().optional(),
+  type: DefinitionPropertyTypeSchema,
+  description: z.string().optional(),
+  group: z.string().optional(),
+  defaultValue: PrimitiveValueSchema.or(ComponentPropertyValueSchema).optional(),
+  validations: z
+    .object({
+      required: z.boolean().optional(),
+      format: z.literal('URL').optional(),
+      in: z
+        .array(
+          z.object({
+            value: z.union([z.string(), z.number()]),
+            displayName: z.string().optional(),
+          }),
+        )
         .optional(),
-    }),
-  ),
+    })
+    .optional(),
+});
+
+export const ComponentVariablesSchema = z.record(
+  z.string().regex(/^[a-zA-Z0-9-_]{1,54}$/), // Here the key is <variableName>_<nanoidId> so we need to allow for a longer length
+  ComponentVariableSchema,
+);
+
+const ComponentSettingsSchema = z.object({
+  variableDefinitions: ComponentVariablesSchema,
 });
 
 const UsedComponentsSchema = z.array(
